# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Margen",
    "description": "Asistente financiero personal. Todos los montos viajan como entero de centavos; el servidor no emite decimales en ningún camino monetario.",
    "version": "v1"
  },
  "paths": {
...
    "/transactions/cash": {
      "post": {
        "tags": [
          "Efectivo"
        ],
        "operationId": "CreateCashTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCashRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/transactions/{id}": {
      "get": {
        "tags": [
          "Movimientos"
        ],
        "operationId": "GetTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Movimientos"
        ],
        "operationId": "PutTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "tags": [
          "Avisos"
        ],
        "operationId": "ListNotifications",
        "parameters": [
          {
            "name": "includeResolved",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/notifications/{id}/resolve": {
      "post": {
        "tags": [
          "Avisos"
        ],
        "operationId": "ResolveNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/emails": {
      "get": {
        "tags": [
          "Correos entrantes"
        ],
        "operationId": "ListIncomingEmails",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateCashRequest": {
        "required": [
          "amountCents",
          "merchant",
          "categoryId",
          "occurredOn",
          "notes"
        ],
        "type": "object",
        "properties": {
          "amountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "merchant": {
            "type": "string"
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "occurredOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PutTransactionRequest": {
        "required": [
          "categoryId",
          "status",
          "notes"
        ],
        "type": "object",
        "properties": {
          "categoryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "status": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "createRule": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
    }
  },
...
}